Implement web storage adapter
This is the first adapter of many to come. The adapter using the initial
state could probably be more simple, but the first idea for the package
was using local/session storage

The reason I renamed the class, is that it was not a manager at all. It
is an adapter, as it adapts the functionality of setting and getting
vales from the web storage.

I should have created CHANGELOG.md a couple of commits ago, but no
problem as there are other key implementions to list ahead.